fix(migrations): cast payment_schedule amount to numeric in 0034 backfill
The grant_payments backfill selected `elem->>'amount'` (text) into the
`amount` numeric NOT NULL column; Postgres won't implicitly cast text→numeric
in INSERT ... SELECT, so the statement failed at parse time (even with zero
matching rows). This is why prod migrate has been stuck since the grants
feature landed — staging had 0034 recorded as applied during development, so
CI never re-executed the SQL there; prod was the first env to actually run it.

Verified against prod in a rolled-back transaction: the full 0033→0036 chain
now applies cleanly (backfill matches 0 rows on prod).

+++ drizzle/migrations/0034_glorious_nighthawk.sql
@@ -13,7 +13,7 @@ WHERE a."status" = 'awarded'
   AND NOT EXISTS (SELECT 1 FROM "grants" g WHERE g."application_id" = a."id");
 --> statement-breakpoint
 INSERT INTO "grant_payments" ("grant_id", "instalment_no", "amount", "due_date")
-SELECT g."id", (elem->>'instalment')::int, elem->>'amount', elem->>'date'
+SELECT g."id", (elem->>'instalment')::int, (elem->>'amount')::numeric, elem->>'date'
 FROM "grants" g
 JOIN "applications" a ON g."application_id" = a."id"
 CROSS JOIN LATERAL jsonb_array_elements(a."payment_schedule") elem
